Xilpm: versal: server: Resolve misra_c_2012_rule_10_4_violation
MISRA:
- Rule 10.4: Both operands of an operator in which the usual
	     arithmetic conversions are performed shall have the same
	     essential type category

To resolve this MISRA violations, do explicit typecast as per requirements.


Acked-for-series: Jolly Shah <jolly.shah@xilinx.com>

+++ lib/sw_services/xilpm/src/versal/server/xpm_device.c
@@ -958,7 +958,7 @@ XStatus XPmDevice_Init(XPm_Device *Device,
 	 * Add requirement by default for PMC subsystem;
 	 * for all the devices except memory region devices.
 	 */
-	if (0U == IS_MEM_REGN(Id)) {
+	if (0U == (u8)IS_MEM_REGN(Id)) {
 		Status = XPmRequirement_Add(
 				XPmSubsystem_GetByIndex((u32)XPM_NODEIDX_SUBSYS_PMC),
 				Device,
